refactor: reorganizacao da branch frontend para o padrao
