+++ frontend/src/assets/hero.png
